shifty: position guessing from name
If shifty.config.guess_position is true (default), when adding or
setting a tag, shifty will check first character of the name and if
it's a number 1-9 that will be used as tag's position. Explicitly
providing 'position' overrides this.

This makes the following presets effectively identical:

 ["1:www"] = { layout = "max", position = 1 },
 ["1:www"] = { layout = "max"               },

 If tag '1:www' was to have default layout as well, you wouldn't even
have to mentione the tag in config.tags, merely config.apps:

 { match = { "Firefox.*" }, tag = "1:www" },

Signed-off-by: koniu <gkusnierz@gmail.com>