shifty: tag indices are now remembered
If config.remember_index is true (default) shifty will keep track of
tag's index (position in tag list) and if closed it will reopen it at
previous index. Position, index and rel_index override this.

This is useful eg.in the following situation:

1. user makes a new tag by pressing mod+t and renames it 'gimp'

2. a config.tags rule creates a new tag for gimp and spawns 'gimp'
   ["gimp"] = { layout = "tile", spawn = "gimp", mwfact = 0.185546 }

3. gimp opens a splash screen and (if not leave_kills) the tag is
destroyed after that closes and before the program window shows up

4. the tag gets re-created -at the same index- when gimp opens the
toolbox (rather than at the end of the tag list).

Signed-off-by: koniu <gkusnierz@gmail.com>