shifty: add tag property 'sweep_delay'
This flag will delay empty tag removal. If after the delay tag has
clients it will not be deleted.

Example:
1. after a crash Firefox shows a dialog asking to restore tabs
2. the dialog creates a new tag (eg. "www" - from config.apps)
3. user closes the dialog
4. "www" tag is closed and switches to the previously selected tag
5. the browser window finally opens once again creating "www" and
   switching to it.

The effect is that when opening a browser, the tag is switched for a
second to the previous one and back to "www". To avoid this we can
delay sweep after the dialog is closed for however long it needs Firefox
to pop up the browser window.

 ["www"] = { exclusive = true, sweep_delay = 1, },

This can be also used for programs that display splash screens causing
the same issue (eg GIMP).

This commit also removed two dead properties: notext (replaced by
icon_only) and skip_taglist (unimplemented).

Signed-off-by: koniu <gkusnierz@gmail.com>