shifty: completion() gets sources, matchers args
This enables use of the completion() function with a limited set of
sources/matchers used for completion matching.

This allows easy creation of custom prompts for use with shifty.

Below example shows an rc.lua entry that shows "Jump to tag" prompt
with completions of currently existing tags.

awful.key({ "Mod5" }, "/", function ()
  awful.prompt.run({ prompt = "Jump to tag: " },
    mypromptbox[mouse.screen].widget,
    function(n)
        local t = shifty.name2tag(n)
        if t then awful.tag.viewonly(t) end
    end,
    function (cmd, cur_pos, ncomp)
        return shifty.completion(cmd, cur_pos, ncomp, { "existing" })
    end)
end)

To only match tag names that begin with the input that's being
completed you can use:

  return shifty.completion(cmd, cur_pos, ncomp, { "existing" }, { "^" })

Signed-off-by: koniu <gkusnierz@gmail.com>