shifty: improvements to per-tag keybindings
With recent changes to awesome we can have more than one function
bound to the same key combo. This interferes with config.tag keys
setting by causing a combo which exist in both globalkeys and tagkeys
to run both the global and tag-related action.

To tackle this we now have a helper function squash_keys() which takes a
table of keyobjects and returns it without duplicated bindings choosing
always the last action defined for a given combo.

As a result, setting a keybinding for a given tag will, for that tag,
override any global binding that uses the same key combination.

In some cases users might want to use the default behavior and overload
a key combo so that in a given tag both - the action specified in
config.tags keys and the global action - are executed. To achieve this
they can use a new config.tags flag - overload_keys = true.

This commit also introduces caching of the keybindings into tag
properties so the list of keybindings doesn't need to be aggregated
on each arrange (ie. tag switch).

Signed-off-by: koniu <gkusnierz@gmail.com>